Rewrite Trie logic in TS
